Backend/database.py
import mysql.connector
from mysql.connector import Error
from config.config import DATABASE_CONFIG

def get_db_connection():
    try:
        conn = mysql.connector.connect(**DATABASE_CONFIG)
        if conn.is_connected():
            return conn
    except Error as e:
        print(f"Error al conectar a la base de datos: {e}")
        return None
